# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "contact": {
      "email": "koldo.urrosolo@opendeusto.es"
    },
    "title": "Main_Server (API-GATEWAY)",
    "description": "API Gateway encargada de la gestión de los usuarios. Además, ofrece la posibilidad de realizar peticiones a las API de equivalencias y de los smartpugs.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8876/"
    }
  ],
...
  "paths": {
...
    "/getSmartplug": {
      "post": {
        "description": "Comprubea si el smartplug está asignado a otro usuario",
        "operationId": "smartplug",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/smartplug"
              }
            }
          },
          "description": "Smartplug location"
        },
        "security": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/smartplugOk"
                }
              }
            },
            "description": "Alias asociado al smartplug"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/smartplugErr"
                }
              }
            },
            "description": "El smartplug lo tiene asignado otro usuario"
          },
          "406": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/smartplugErr"
                }
              }
            },
            "description": "No se ha encontrado el smartplug"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/smartplugSwitch": {
      "get": {
        "description": "Enciende o apaga el smartplug en función de su estado actual",
        "operationId": "smartplugSwitch",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            },
            "description": "Mensaje de estado"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/smartplugNoEncontrado"
                }
              }
            },
            "description": "No se ha encontrado el smartplug"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            },
            "description": "Respuesta por defecto"
          }
        },
        "summary": "Enciende el smartplug",
        "tags": [
          "External"
        ]
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "smartplugOk": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "alias": {
            "type": "string"
          }
        }
      },
      "smartplugErr": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "err": {
            "type": "string"
          }
        }
      },
...
      "success": {
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "smartplugNoEncontrado": {
        "additionalProperties": false,
        "properties": {
          "msg": {
            "description": "Smartplug no se ha encontrado",
            "maxLength": 29,
            "minLength": 29,
            "pattern": "[a-zA-Z\\s]+",
            "type": "string"
          }
        },
        "readOnly": true,
        "type": "object"
      },
      "smartplug": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Ubicación donde se encuentra el smartplug. Tiene el siguiente formato: sp_w.r{Fila}.c{Columna} *Hay que sustituir el valor de Fila y Columna por el correspondiente.\n",
            "maxLength": 12,
            "pattern": "sp_w.r\\d{1,2}.c\\d{1,2}",
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  }
}